+++ docs/.vuepress/config.js
@@ -197,7 +197,12 @@ module.exports = {
         {
           title: "How-To",
           collapsable: false,
-          children: listDir("how-to", ["yaml-media", "middleware", "tortoise"])
+          children: listDir("how-to", [
+            "yaml-media",
+            "middleware",
+            "tortoise",
+            "socketio"
+          ])
         }
       ],
       "/discussions/": [