+++ tests/conftest.py
@@ -5,18 +5,26 @@ from typing import NamedTuple
 
 import pytest
 
-from bocadillo import API, App, Recipe, Templates
+from bocadillo import App, API, Templates, Recipe
+from bocadillo.testing import create_client
 
 
 # Tests that use the `app` fixture will run once for each of these
 # application classes.
 APP_CLASSES = [App, API, lambda: Recipe("tacos")]
+CLIENT_FACTORIES = [create_client, lambda app: app.client]
 
 
 @pytest.fixture(params=APP_CLASSES, name="app")
 def fixture_app(request):
     cls = request.param
     return cls()
+
+
+@pytest.fixture(params=CLIENT_FACTORIES)
+def client(app, request):
+    factory = request.param
+    return factory(app)
 
 
 @pytest.fixture(name="templates")