+++ tests/test_hsts.py
@@ -1,4 +1,5 @@
 from bocadillo import App
+from bocadillo.testing import create_client
 
 
 def test_if_hsts_enabled_and_request_is_on_http_then_redirects_to_https():
@@ -8,6 +9,7 @@ def test_if_hsts_enabled_and_request_is_on_http_then_redirects_to_https():
     async def index(req, res):
         pass
 
-    response = app.client.get("/", allow_redirects=False)
+    client = create_client(app)
+    response = client.get("/", allow_redirects=False)
     assert response.status_code == 301
     assert response.headers["location"] == "https://testserver/"